+++ src/unnest.cpp
@@ -1,13 +1,28 @@
 #include "unnest.h"
 
+Unnester::ProcessUnnamed sexp2unnamed(SEXP x) {
+  if (x == R_NilValue)
+    return Unnester::ProcessUnnamed::NONE;
+  if (TYPEOF(x) == STRSXP) {
+    const char* nm = CHAR(STRING_ELT(x, 0));
+    if (!strcmp(nm, "stack"))
+      return Unnester::ProcessUnnamed::STACK;
+    else if (!strcmp(nm, "exclude"))
+      return Unnester::ProcessUnnamed::EXCLUDE;
+    else if (!strcmp(nm, "as_is") || !strcmp(nm, "as.is") || !strcmp(nm, "asis"))
+      return Unnester::ProcessUnnamed::ASIS;
+    else if (!strcmp(nm, "paste"))
+      return Unnester::ProcessUnnamed::PASTE;
+  }
+  Rf_error("Invalid `unnamed_list` argument. Must be one of 'as_is', 'exclude', 'stack', 'paste' or NULL");
+}
+
 // simple stacker
 void Unnester::stack_nodes(NodeAccumulator& acc, VarAccumulator& vacc,
                            const Spec& pspec, const Spec& spec,
                            uint_fast32_t ix, const vector<SpecMatch>& matches,
-                           const bool rep_to_max = false) {
+                           const bool stack_atomic = false) {
   P(">>> stack_nodes ---\n");
-  size_t N = matches.size();
-
   R_xlen_t beg = 0, end=0;
   unordered_map<uint_fast32_t, unique_ptr<RangeNode>> out_nodes;
 
@@ -18,12 +33,11 @@ void Unnester::stack_nodes(NodeAccumulator& acc, VarAccumulator& vacc,
 
   uint_fast32_t cix = child_ix(ix, spec.name);
 
-  int i = 1;
   for (const SpecMatch& m: matches) {
     NodeAccumulator iacc;
     VarAccumulator ivacc(vacc.dedupe);
 
-    dispatch_match_to_child(iacc, ivacc, pspec, spec, cix, m);
+    dispatch_match_to_child(iacc, ivacc, pspec, spec, cix, m, stack_atomic);
     end += iacc.nrows;
 
     // add index
@@ -65,7 +79,7 @@ void Unnester::stack_nodes(NodeAccumulator& acc, VarAccumulator& vacc,
     acc.pnodes.push_front(move(onode.second));
   }
 
-  if (rep_to_max || this->rep_to_max)
+  if (stack_atomic || this->rep_to_max)
     acc.nrows = max(acc.nrows, end);
   else
     acc.nrows = acc.nrows * end;
@@ -78,13 +92,13 @@ void Unnester::stack_nodes(NodeAccumulator& acc, VarAccumulator& vacc,
 void Unnester::stack_nodes(vector<NodeAccumulator>& accs, VarAccumulator& vacc,
                            const Spec& pspec, const Spec& spec,
                            uint_fast32_t ix, const vector<SpecMatch>& matches,
-                           const bool rep_to_max = false) {
+                           const bool stack_atomic = false) {
 
   if (accs.size() == 0) return;
 
   P(">>> gstack_nodes ---\n");
 
-  size_t Ngr = spec.groups.size(), N = matches.size();;
+  size_t Ngr = spec.groups.size();
   if (accs.size() != Ngr)
     Rf_error("Internal: Invalid grouped stack. Accumulator size (%ld) and spec size (%l) mismatch.",
              accs.size(), Ngr);
@@ -161,7 +175,7 @@ void Unnester::stack_nodes(vector<NodeAccumulator>& accs, VarAccumulator& vacc,
       accs[ci].pnodes.push_front(move(onode.second));
     }
 
-    if (rep_to_max || this->rep_to_max)
+    if (stack_atomic || this->rep_to_max)
       accs[ci].nrows = max(accs[ci].nrows, end[ci]);
     else
       accs[ci].nrows = accs[ci].nrows * end[ci];
@@ -170,8 +184,10 @@ void Unnester::stack_nodes(vector<NodeAccumulator>& accs, VarAccumulator& vacc,
 
 }
 
-extern "C" SEXP C_unnest(SEXP x, SEXP lspec, SEXP dedupe, SEXP stack_atomic, SEXP cross_join) {
-  SEXPTYPE type = TYPEOF(x);
+extern "C" SEXP C_unnest(SEXP x, SEXP lspec, SEXP dedupe,
+                         SEXP stack_atomic, SEXP process_atomic,
+                         SEXP process_unnamed_list,
+                         SEXP cross_join) {
   if (TYPEOF(x) != VECSXP) {
 	Rf_error("x must be a list vector");
   }
@@ -179,6 +195,9 @@ extern "C" SEXP C_unnest(SEXP x, SEXP lspec, SEXP dedupe, SEXP stack_atomic, SEX
   Unnester unnester;
   unnester.dedupe = sexp2bool(dedupe);
   unnester.stack_atomic = sexp2bool(stack_atomic);
+  unnester.stack_atomic_df = stack_atomic == R_NilValue;
+  unnester.process_atomic = sexp2process(process_atomic);
+  unnester.process_unnamed_list = sexp2unnamed(process_unnamed_list);
   unnester.rep_to_max = !sexp2bool(cross_join);
 
   return unnester.process(x, lspec);