+++ NEWS.md
@@ -1,5 +1,11 @@
+Version 0.0.3
+=============
+
+- New `default` argument in `s()` specification.
+- Data frame columns are now stacked by default (`stack_atomic = NULL` parameter)
+- New `unnest()` parameter `process_unnamed_lists` for controlling processing of unnamed lists.
 
 Version 0.0.2
 =============
 
-Original CRAN release
+- Original CRAN release